fix(#1078): move materialisation.py from core/ to engine/ layer
First CI run on PR #1112 surfaced two architecture failures:

  Architecture Tests (tests/architecture/test_layer_deps.py)
    Layer 'core/' has forbidden imports:
      core/materialisation.py: imports scieasy.blocks.base.config
      core/materialisation.py: imports scieasy.blocks.registry

  Import Contracts (importlinter)
    Core must not depend on blocks, engine, api, ai, or workflow
      scieasy.core.materialisation -> scieasy.blocks.base.config
      scieasy.core.materialisation -> scieasy.blocks.registry

The helpers fundamentally have to consume BlockRegistry and BlockConfig
from blocks/ to satisfy the planning doc's contract (find_saver /
find_loader dispatch). core/ (Layer 1) cannot import blocks/ (Layer 2).
The issue title actually leads with 'Engine:' — the body's core/ path
was a planning typo.

Files moved (no public API change):
  src/scieasy/core/materialisation.py  -> src/scieasy/engine/materialisation.py
  tests/core/test_materialisation.py   -> tests/engine/test_materialisation.py

Updated:
  - utils/fs.py docstring reference (core.materialisation -> engine.materialisation).
  - CHANGELOG entry: replaced core/ with engine/ and noted the rationale.
  - Impl record: added a 'Layer placement note' explaining the rationale.
  - Test import updated: scieasy.engine.materialisation.

Architecture + engine + utils + registry test suites green locally.
