+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 
-- [#1078] Add `core/materialisation.py` (`materialise_to_file` / `reconstruct_from_file`) plus `utils/fs.py::mount_pathlike` cross-platform link helper (ADR-028 §D8). The two new helpers consume `BlockRegistry.find_saver` / `find_loader` from #1077 to give engine + bridge code a single canonical surface for writing a `DataObject` to a file in a chosen format and constructing a typed `DataObject` from a path; AppBlock prepare/restore follow-ups (#1079 / #1080) will route through these instead of reimplementing format dispatch. `materialise_to_file` prefers a native link (`mount_pathlike`) when `obj.storage_ref` already lives on disk in the target format and falls back to a saver round-trip on link failure; `reconstruct_from_file` falls back to `Artifact(file_path=path, ...)` when no loader matches and `target_type` is an `Artifact` subclass, raises `LookupError` otherwise, and has a documented dynamic-port second-pass (`find_loader(DataObject, ext)` + `core_type` injection) so the core `LoadData` block is reachable for concrete core target types. `mount_pathlike` uses POSIX symlinks on \*nix and prefers Windows directory junctions (no admin) or NTFS hardlinks for files; raises `OSError` when no primitive succeeds. 12 new tests in `tests/core/test_materialisation.py` + 6 in `tests/utils/test_fs.py` (platform-gated symlink / junction / hardlink cases). 688 tests pass across `tests/core` + `tests/utils` + `tests/blocks/test_registry*`. (@claude, 2026-05-18, branch: fix/issue-1078/core-materialisation-helpers, session: 20260518-025336-fix-1078-core-materialisation-py-utils-f)
+- [#1078] Add `engine/materialisation.py` (`materialise_to_file` / `reconstruct_from_file`) plus `utils/fs.py::mount_pathlike` cross-platform link helper (ADR-028 §D8). Module lives at the `engine/` layer rather than `core/` (planning doc said `core/`, but layer-dependency tests forbid `core/` from importing `blocks/` and the helpers must consume `BlockRegistry`; the issue title's "Engine:" prefix is the intended layer). The two new helpers consume `BlockRegistry.find_saver` / `find_loader` from #1077 to give engine + bridge code a single canonical surface for writing a `DataObject` to a file in a chosen format and constructing a typed `DataObject` from a path; AppBlock prepare/restore follow-ups (#1079 / #1080) will route through these instead of reimplementing format dispatch. `materialise_to_file` prefers a native link (`mount_pathlike`) when `obj.storage_ref` already lives on disk in the target format and falls back to a saver round-trip on link failure; `reconstruct_from_file` falls back to `Artifact(file_path=path, ...)` when no loader matches and `target_type` is an `Artifact` subclass, raises `LookupError` otherwise, and has a documented dynamic-port second-pass (`find_loader(DataObject, ext)` + `core_type` injection) so the core `LoadData` block is reachable for concrete core target types. `mount_pathlike` uses POSIX symlinks on \*nix and prefers Windows directory junctions (no admin) or NTFS hardlinks for files; raises `OSError` when no primitive succeeds. 12 new tests in `tests/core/test_materialisation.py` + 6 in `tests/utils/test_fs.py` (platform-gated symlink / junction / hardlink cases). 688 tests pass across `tests/core` + `tests/utils` + `tests/blocks/test_registry*`. (@claude, 2026-05-18, branch: fix/issue-1078/core-materialisation-helpers, session: 20260518-025336-fix-1078-core-materialisation-py-utils-f)
 - [#1075] Declare `supported_extensions` ClassVar on imaging plugin IO blocks `LoadImage` and `SaveImage` (ADR-028 §D8). Removed module-level legacy constants `_TIFF_EXTS` / `_ZARR_EXTS` / `_SUPPORTED_EXTS` (load_image.py) and `_TIFF_FORMAT` / `_ZARR_FORMAT` / `_SUPPORTED_FORMATS` / `_EXT_TO_FORMAT` (save_image.py); routed format dispatch through `self._detect_format(path)`. Both ClassVars expose `{".tif": "tiff", ".tiff": "tiff", ".zarr": "zarr"}` (exact mirror for round-trip discoverability). 16 new `TestSupportedExtensionsClassVar` cases (8 LoadImage + 8 SaveImage) including explicit removed-constants assertions, mirror equality, and TIFF + Zarr smoke round-trips through the new dispatch path; lint clean. (@claude, 2026-05-18, branch: fix/issue-1075/load-save-image-supported-extensions, session: 20260518-022121-fix-1075-declare-supported-extensions-on)
 - [#1074] Declare `supported_extensions` ClassVar on `LoadData` and `SaveData` (ADR-028 §D8). Each block now publishes the union of its per-core-type dispatch suffixes (`.npy`/`.npz`/`.zarr`/`.parquet`/`.pq`/`.pkl`/`.pickle`/`.csv`/`.tsv`/`.json`/`.txt`/`.log`/`.md`/`.html`/`.xml`/`.yaml`/`.yml`/`.toml`) keyed to stable format identifiers. SaveData mirrors LoadData exactly so round-trip discoverability holds; module-level `_resolve_format` / `_resolve_save_format` helpers route per-core-type dispatch through `IOBlock._detect_format` semantics. `ValueError` messages on unknown extensions now derive the supported list from `sorted(self.supported_extensions.keys())` instead of hardcoded substrings. 18 new `TestSupportedExtensionsClassVar` cases (9 LoadData + 9 SaveData) including the explicit mirror-equality assertion; full IO test suite (153/153) green. (@claude, 2026-05-18, branch: fix/issue-1074/load-save-data-supported-extensions, session: 20260518-020747-fix-1074-declare-supported-extensions-on)
 - [#1077] Add `BlockRegistry.find_loader(dtype, ext)`, `find_saver(dtype, ext)`, and `find_io_blocks_for_type(dtype, direction)` query methods (ADR-028 §D8). Each method consults the cached `supported_extensions` ClassVar (added on `BlockSpec` so dispatch does not require re-importing the block class) and uses `TypeSignature.matches` (ADR-027) for type compatibility — loaders are filtered by output-type-IS-A-dtype, savers by dtype-IS-A-saver-accepted-type (the classical covariance/contravariance flip). Disambiguation prefers the longest matching accepted-type chain (most specific); first-registered-wins on ties. Class resolution mirrors `instantiate()`: mtime-keyed reload for Tier-1 drop-ins, regular `import_module` otherwise; per-block import failures drop the candidate silently to match scan-loop tolerance. 17 new tests in `tests/blocks/test_registry_find.py` covering exact match, subtype direction (both polarities), specificity ordering, first-registered tie-break, case-insensitivity, `dtype=None` shortcut, enumeration order, and invalid-direction error. Full registry suite (77) + architecture suite (273) green. (@claude, 2026-05-18, branch: fix/issue-1077/blockregistry-find-loader-saver, session: 20260518-021831-fix-1077-blockregistry-find-loader-find)
