+++ docs/audit/impl-records/phase-minus-1/issue-1078-33312d0.md
@@ -1,8 +1,23 @@
 # Implementation record: #1078
 
+## Layer placement note
+
+The planning doc and issue body named the new module
+``src/scieasy/core/materialisation.py``. The first CI run on PR #1112
+surfaced `tests/architecture/test_layer_deps.py` and the
+`importlinter` import-contract test rejecting the file because
+``core/`` (Layer 1) is forbidden from importing ``blocks/`` (Layer 2),
+and the helpers fundamentally have to consume `BlockRegistry` and
+`BlockConfig` from `blocks/`. The module was moved to
+``src/scieasy/engine/materialisation.py`` (Layer 3, allowed to import
+both ``core/`` and ``blocks/``) — the issue title actually leads with
+"Engine:" so this is the intended layer; the body's ``core/`` path was
+a planning typo. Test file moved in lockstep:
+``tests/engine/test_materialisation.py``. No public API change.
+
 ## Files modified
 
-- `src/scieasy/core/materialisation.py` (NEW) — public helpers
+- `src/scieasy/engine/materialisation.py` (NEW) — public helpers
   `materialise_to_file(obj, dest_dir, extension=None, *, filename_stem,
   registry) -> Path` and `reconstruct_from_file(path, target_type,
   extension=None, *, registry) -> DataObject`. Both consult
@@ -18,7 +33,7 @@
   (POSIX symlink, Windows directory junction via `_winapi.CreateJunction`,
   Windows file hardlink via `os.link`). Raises `OSError` when no
   primitive succeeds so callers can fall back to a byte copy.
-- `tests/core/test_materialisation.py` (NEW) — 12 tests covering
+- `tests/engine/test_materialisation.py` (NEW) — 12 tests covering
   round-trip DataFrame (.csv) and Array (.npy), Artifact fallback,
   LookupError for concrete-type miss, missing-file error,
   pass-through invocation (with mock + side-effect to satisfy the
@@ -93,7 +108,7 @@
 
 ## Tests added
 
-- `tests/core/test_materialisation.py`
+- `tests/engine/test_materialisation.py`
   - `test_round_trip_dataframe` — DataFrame → CSV → DataFrame.
   - `test_round_trip_array` — numpy Array → .npy → numpy Array,
     asserts content equality.
